+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.30] - 2021-07-09
 
+### Update
+
+*[Bump commander@8.0.0.](https://github.com/tj/commander.js/releases/tag/v8.0.0)
+
 ## [0.0.29] - 2021-06-29
 
 ### Added