+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Update
 
-*[Bump commander@8.0.0.](https://github.com/tj/commander.js/releases/tag/v8.0.0)
+* [Bump commander@8.0.0](https://github.com/tj/commander.js/releases/tag/v8.0.0)
+* [Bump @restqa/cucumber-export@0.1.10](https://github.com/restqa/cucumber-export/releases/tag/0.1.10)
 
 ## [0.0.29] - 2021-06-29