+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.30] - 2021-07-09
 
+### Added
+
+* Enable new command to mange the telemtry : `restqa telemetry off` | [#119](https://github.com/restqa/restqa/pull/119)
+
 ### Update
 
+<<<<<<< HEAD
 * [Bump commander@8.0.0](https://github.com/tj/commander.js/releases/tag/v8.0.0)
 * [Bump @restqa/cucumber-export@0.1.10](https://github.com/restqa/cucumber-export/releases/tag/0.1.10)
 * [Bump @restqa/restqapi@0.0.19](https://github.com/restqa/restqapi/releases/tag/0.0.19)